Mark ForestOptions input and beta (#25308)
## Description

Mark ForestOptions  as input and beta.

Note that this should have no impact on what beta users can do currently
for two reasons:

- None of the ForestTypes are beta, so you can't actually use it to opt
into anything unless also importing an alpha ForestType
- None of the APIs which accept ForestOptions are beta, so there is
nothing to do with it.

This is still a useful change as it is progress toward making some of
the APIs which can optionally include ForestOptions (like
IndependentView and configuredSharedTree) available as beta, even if
users of them won't be able to customize which forest is used (as the
ForestType options are alpha).

Note that ForestType is a sealed opaque type, so stabilizing this really
only pins down its name and semantics, not anything about the
implementation.