Add FormatValidator to replace JsonValidator (#25311)
## Description

This is intended as work toward stabilizing ICodecOptions as beta
without committing to details of how we validate data.

Doing this as a two part change (with removal of JsonValidator from
ICodecOptions, and moving ICodecOptions to beta would be next sometime
after the next release) will better accommodate the existing use of
these APIs in at least one known app.