Make presence tests use standard base config, fixing test result reporting (#25366)
## Description

Presence has 1 pending and 132 passing tests.

Since it runs all these tests in both cjs and esm mode, that's a total
of 2 pending and 164 passing tests which should be added to the results.

Unfortunately, the pipeline test dashboard only shows an additional 1
pending and 132 passing tests since only one copy of the tests is
correctly reported: this is not an issuer specific to presence but
applies to all packages which have both kinds of tests.

Total tests are now 36317, and baseline was 36184 according to the test
reports (on GitHub and ADO)