improvement(task-manager): Ignore stashed ops (#25387)
## Description

This PR updates TaskManager to no longer process stashed ops. This is
because any ops that were submitted are no longer "valid" due to
TaskManager limitations around offline behavior. This is similar to the
resubmit scenario (TaskManager does not resubmit ops at all). For each
op, here is the justification as to why we shouldn't process stashed
ops:

- `volunteer` - Going offline while waiting for a task assignment will
cause the volunteer promise to automatically `reject`. Similar to
reconnect scenarios, we should let the app developer re-volunteer in
this scenario.
- `abandon`/`complete` - These ops only makes sense to apply for tasks
that the client is assigned to. Since going offline will lose any task
assignments, we can safely ignore any abandon ops.

## Misc

[AB#44704](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/44704)
Fixes
[AB#5341](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5341)
Fixes
[AB#3985](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3985)