Only generate response stacks when no stack exists (#25312)
The code to generate stacks on responses had some performance issue.
Specifically, the code would generate a stack even when it wasn't
necessary. This change fixes those issues to first checks if a stack
exists and only creates a stack when one does not exist.

This only has a minor impact on production scenarios where errors are
uncommon, but in tests, like the local server stress tests, we can hit
this case a lot which slows down the test runs overall.