Refactor how schema statics work (#25444)
## Description

Tweak how schema statics are setup on SchemFactory.

This has been confirmed to improve IntelliSense, since it now points to
SchmeaStatics where the actual docs are rather than showing the
inheritDoc comments.

It may somewhat degrade the experience on the website, but the correct
docs should still be reachable, and future versions of the site might
inline the inherited members.

Applying this pattern to ShcmeaFactoryAlpha was unsuccessful, but even
just applying it to the public API surface saves a lot of code and
improves IntelliSense.