Smoke test tree CJS (#25467)
## Description

Introduce the idea of smoke tests in tree, and use it to enable a single
test to run on CI for the CJS built (instead of none).

As appropriate tests are found, more can be opted in.

Including startup, running this smoke test suite takes under 2 seconds,
so it should not significantly regress CI time.

This pattern is planned to also support prod mode added in
https://github.com/microsoft/FluidFramework/pull/24013 so it can have at
least some minimal coverage on CI as well.

This will guard again having test configs which are full broken (like
happened with experimental/tree's ESM build recently) as well as help
detect if the package itself has major issues in the smoke tested
config.

Note that due to the fact that the CI test results viewer does not fails
to handle multiple tests runs from a single package, resulting in the new run not displaying.

I can confirm however that the code coverage from both runs is getting
merged as it did not regress and instead covered the same lines, but
found a few new branches (which were covered), likely due to some
oddities of how the branch coverage is implemented and/or how CJS was
compiled (for example one of the changes was binding two branches in a
line which was just a function declaration, and another was in an
index.ts file.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).