fix(api-markdown-documenter): Ensure no line breaks in HTML table cell content (#25497)
Fixed an issue where HTML table cell content was generated with line
breaks for formatting. This would break Markdown table syntax. Line
breaks are now omitted in this context.

I have been unable to reproduce this behavior outside of the context of
our fluidframework website build (even passing the same problematic
content from that build into tests within the documenter library). It
isn't clear to me what is causing the line breaking behavior in one
setting and not the other. But I have verified that this fixes the
behavior for the website build.