feat(eslint-config-fluid): Enable new TypeScript rules as warnings (#25508)
`@typescript-eslint/eslint-plugin` has deprecated and disabled the
following rules:

- [ban-types](https://typescript-eslint.io/rules/ban-types/)
-
[no-empty-interface](https://typescript-eslint.io/rules/no-empty-interface/)

They are replaced by the following new rules, which are now configured
as errors in this library:

-
[no-empty-object-type](https://typescript-eslint.io/rules/no-empty-object-type/)
-
[no-unsafe-function-type](https://typescript-eslint.io/rules/no-unsafe-function-type/)
-
[no-wrapper-object-types](https://typescript-eslint.io/rules/no-wrapper-object-types/)

This PR pre-emptively updates our configs to set these new rules as
warnings. In the next major version bump, the old rules will be removed
and these new rules will be promoted to errors.