docs(tree): general schema evolution docs (#25218)
## Description

- adds a landing page for schema evolution topics in fluid docs with
general information
- adds page on types of schema changes and how upgrading them works
- removes doc from shared tree package

## Reviewer Guidance

This consolidates the docs on schema evolution we had stored in a few
different locations and attempts to simplify them for users of
SharedTree, leaving out a lot of the more conceptual information on
tools that _could_ be implemented to make schema evolution better and
just focusing on what can be used.

Some of the docs are existing and just copied from other places but
everything should be reviewed in case we can improve them. In
particular, the example code in the sample policies page should be
reviewed.

Staged allowed types are only briefly mentioned but will get more
comprehensive docs/linking in a separate pr.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>
Co-authored-by: Noah Encke <78610362+noencke@users.noreply.github.com>