Migrate off of process global and openStdin (#25569)
## Description

Based on
https://stackoverflow.com/questions/33831978/unknown-method-process-openstdin#comment126485133_33832075
appears we do not need to resume the stream unless using "old" streams
mode (from back in NodeJS 0.10), and can just use process.stdin instead
of openStdin which was [as done in Node 0.3.3 back in
2011](https://github.com/nodejs/node/blob/43e4efdf210adb2cc3ba26518fd4588f9e0152ff/doc/changelogs/CHANGELOG_ARCHIVE.md#20110102-version-033-unstable).

CheckPolicy's `--stdin` flag is the older user of this code, and was
manually tested to work. As far as I can tell nothing uses this feature.