fix(tree): handle redundant sequencing of local commit (#25544)
## Description

Fixes a bug in the scenario where a commit from a child shared branch is
concurrently merged into the main shared branch by different peers. The
peer whose merged commits are not sequenced first had a mistaken
expectation that it should have a commit on its main local branch for
each commit that is sequenced by the service if that commit came from
this peer. This is no longer a valid expectation.

## Breaking Changes

None