refactor(tree): make codec dependencies more explicit and auditable (#25619)
## Description

The main goal of this PR to make the hierarchy of codec dependencies
that are scattered across the SharedTree codebase more explicit and more
auditable. This is accomplished in two ways:

### Specific Version Types

Specific types are introduced for each format version and used whenever
possible. For example, referring to a specific version of the
`EditManager` format is now done with an instance of
`EditManagerFormatVersion` instead of a plain `number` or
`FormatVersion`.

This makes clearer what the space of version number is being addressed.
The type acts as a code reference that can be followed with Ctrl+Click.

This also enforces that version literals correspond to a valid version
for the target version space. For example, the code `const
editManagerVersion: EditManagerFormatVersion = 42;` will not compile
because there is no version 42 for the edit manager format.

### Queryable Codec Trees

This PR introduces the `CodecTree` type so that dependencies between
codec versions can be reified. Most codecs involved in SharedTree now
come with an associated free function that returns a `CodecTree` for a
given version of that codec.
These are used in a new set of snapshot tests that capture the codec
dependency trees of each SharedTree format version. While the primary
motivation of such tests is to end up with files that can easily be
inspected, these tests also have value in preventing unintentional
changes in codec dependencies.

## Other Changes

The changes involved in building the above have also led to the
realization that some dependencies were hard coded in the wrong place:
the format version for representing SharedTree changes used to be hard
coded in the `EditManager` and `Message` codecs. This was questionable
because `EditManager` and `Message` are completely agnostic to such
details. These dependencies are now hard-coded in sharedTree.ts.

## Breaking Changes

None