fix(directory): Relax pending entry bookkeeping when rolling back subdirectory ops (#25598)
## Description

This PR relaxes the `0xc37` and `0xc38` assert conditions when trying to
rollback subdirectory ops.

There is an edge case that was surfaced in local stress test seed #124
which covers the following scenario:
1. Client A creates subdirectory `/x` and subdirectory `/x/y` (pending,
unack'd)
2. Client A enters staging mode
3. Client A deletes subdirectory `y` (in staging mode)
4. Client B creates `/x` (pending, unack'd)
5. Client B deletes `/x` (pending, unack'd)
6. Client C creates `/x` (pending, unack'd)
7. All non-staging mode ops are ack'd and processed
8. Client A exits staging mode and does not commit

In this scenario, when Client A exits staging modes and tries to
rollback all the op sent in step (3), it is able to successfully find
the "correct" subdirectory to rollback this op for (since subdirectory
`/x` was re-created by Client C in step (6)). However, since the
subdirectory was deleted and re-created by remote ops, the pending data
for the delete subdirectory created in step (3) was lost, leading to the
assert being hit since it cannot find the corresponding pending data.

There is similar logic for other rollback operations well. In
non-rollback scenarios, we rely on the incoming message's sequence data
to determine if an op is targeting the current instance of a
subdirectory (`isMessageForCurrentInstanceOfSubDirectory()`). However,
since we only have a local unack'd op, this data does not exist.

## Misc

[AB#48147](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/48147)