Improve Assert APIs (#24013)
## Description

Allows passing debug/development only formatted text into asserts via
callbacks.
This change makes it possible to have descriptive assert messages with
string interpolation for debug/development builds, but still get the
benefits of minimized tagged asserts in production.

Enables debugAsserts by default in development configurations, and adds
a more robust option for testing with them (and similar development
only) code disabled.

The main risk associated with this change is that someone might
accidently do all of the following:
1. Add a debugAsserts condition which is impure (has side effects) (The
documentation for debugAssert explicitly says not to do that, but people
can have bugs)
2. Some important scenario fails to work correctly (due to a second
bug), but happens to be fixed by the side-effect of the debugAssert.
3. None of the mechanisms for testing the production mode of this code
are used.

This PR includes examples of three approaches for catching such issues:

1. Cases where debugAssert is more heavily used can have test suites
which explicitly disable it (or loop over both modes). Since debugAssert
is often most useful in performance sensitive code, doing this for the
performance test suite is a good option. Example in [jsonCursor.bench.ts

](https://github.com/microsoft/FluidFramework/pull/24013/files#diff-039f1dc67b55f6906f5afb4238afa30210646fab1a542fcb8c0beca1646ad73a)
2. A package level script can be added to run the all the tests in prod
mode, see
[test:mocha:prod](https://github.com/microsoft/FluidFramework/pull/24013/files#diff-34daa9b8b6fa1ce21c7854703df1f12799bbc4aa32e0bc8363418bc9df10a356R136)
3. The above script can be fully or partly included in the CI tests, see
[test:mocha](https://github.com/microsoft/FluidFramework/pull/24013/files#diff-34daa9b8b6fa1ce21c7854703df1f12799bbc4aa32e0bc8363418bc9df10a356R133)
which runs the smoke test subset.

These costs need to be weighed against the benefits:

1. There is now a way to do test/development only validation, which
means from bundle size and performance perspective more validation is
acceptable to add which can detect bugs during development and testing
which would have gone unnoticed.
2. Some validation which was only useful for development and testing can
be moved to the new API, improving performance and bundle size.
3. Error cases which would really benefit from having some formatted
messages no longer have to choose between regressing bundle size (by not
being a tagged assert) OR having poor message quality during
development. Both the debugAssert and the new debug messages for normal
assert fill this gap.

Adoption of these features can start in Tree where testing is setup for
it, and which is particularly sensitive to bundle size and performance
and does a large amount of validation. With the changes in this PR
alone, it has already detected on bug (in a test), improved performance
and should reduce bundle size.


## Breaking Changes

The format for assertion error messages in builds that do not enable
production bundle optimizations can now include a second line with
additional text.

This could presumably break some things, but it is opt in and has no
impact on existing asserts or behavior of production bundles.