Move tree-agent plugins to packages (#25659)
This moves the langchain-backed `SharedTreeChatModel` plugin and the
SES-backed edit evaluator to seperate packages so that they will not be
downloaded by consumers who are not using them. Integration tests that
rely on langchain have also moved into the langchain package. This also
removes the deprecated `createSemanticAgent` helper and makes some small
fixups to tree-agent.