build(eslint-config-fluid): migrate from eslint-plugin-import to eslint-plugin-import-x (#25661)
Migrates to eslint-plugin-import-x as eslint-plugin-import is no longer
maintained.
Updated all references in config files and dependencies.

Once this is released, integration into main will require renaming
comments and overrides, but the changes will be mechanical.

This is technically a breaking change, so I bumped it to a major.
Regardless, I want to integrate this change alone, so I want to release
once this is merged. The only question is if this needs to be a major
release. If so, that's fine, but several of the subsequent changes will
also be "breaking" so we'll be releasing several majors in short
succession. I think that's fine.