improvement(tree): Add index-based get method overloads to TableSchema APIs (#25692)
Adds support for index parameters to the following methods on `Table`
schema:
- `getColumn`
- `getRow`
- `getCell`
- `setCell`

Also fixes a bug where column/row nodes provided to these (and other)
table methods would succeed if the node's ID matched that of a node in
the table. These cases will now fail if the node object is not the same.