(tree) Incremental summary: Fix tracking state not recreated during load and other bugs (#25421)
This PR fixes the following bugs in incremental summary of tree:
- When the forest summarizer is loaded, the chunk tracking state was not
recreated. This means that when a new summarizer client is chosen, it
will start from an empty state and resummarize everything.
This is fixed by updating the `getEncodedIncrementalChunk` function to
`decodeIncrementalChunk`. It now takes a callback function as parameter
which decodes the encoded chunk contents into a tree chunk. The
`ForestIncrementalSummaryBuilder` uses this chunk and the data it tracks
during loading to recreate the tracking state.
- The chunk reference ID always starts at 0 in a new client. This means
that after forest summarizer loads, if at a given depth in the
incremental tree, a subtree with ID 0 already existed, a new subtree
could get the ID 0 as well. The same goes for any other ID.
This is fixed by updating the `nextReferenceId` to the next value after
the max reference ID encountered during load.

Also, added tests to validate these scenarios.


[AB#47252](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/47252)