test(dataStoreRuntime): Add test coverage for error string since it's the only way to detect a given internal case (#25711)
A partner has a new need to detect if a channel is not found. See
Background below. Today this is not possible apart from inspecting the
error message. So this change adds test coverage for that string so we
will think twice before changing it.

This is a temporary stop-gap.

### Background

Now that we are supporting a so-called "MultiFormatDataObject" where we
have to probe for various possible root channels, we need to be able to
find out that a channel doesn't exist.

Today, `FluidDataStoreRuntime.getChannel` throws a basic error in this
case, but we need to distinguish this case from other errors thrown.