Add `isArray` and `isMap` to tree-agent context (#25725)
This gives the LLM an alternative to `Array.isArray` and `instanceof
Map` that will work on our node types, since `Array.isArray` does not
work with our non-pojo array classes.

The provided functions also work for JS arrays and Maps, so the LLM
doesn't have to distinguish between when to use the built-in
Array.isArray vs. when to use our provided `isArray` helper - it can
always use the latter. This also gives us the option in the future to
force it to use ours by making `Array.isArray` error within the
boundaries of the eval.

This also adds some explicit tests in SharedTree for `Array.isArray`
support.