(tree) Add incremental summary optimzation opt-in to AnnotatedAllowedTypes (#25408)
## Description
With this PR, a set of allowed types in a schema can be opted in to
incremental summary optimization. These allowed types will be optimized
during summary such that if they don't change across summaries, they
will not be encoded and their content will not be included in the
summary that is uploaded to the service. The usage pattern is described
below.

## Usage pattern
- Incremental summarization only works for forest type
`ForestTypeOptimized` when the compression strategy
is `TreeCompressionStrategyExtended.CompressedIncremental`. In addition,
`SharedTreeOptionsInternal.shouldEncodeFieldIncrementally` must be
passed when creating the tree. This callback function will be called for
each allowed types in the schema to determine if it should be
incrementally summarized. These configs can be passed in when creating a
tree via the `configuredSharedTree` API.
- A set of allowed types can be opted in to incremental summary
optimization by adding an `incrementalSummaryHint`symbol as true to the
`AllowedTypesMetadata.custom` property.
- The helper function `shouldIncrementallySummarizeAllowedTypes` (added
in this PR) can be used to implement the
`SharedTreeOptionsInternal.shouldEncodeFieldIncrementally` callback
function. It takes in the schema that contains the allowed types and the
node identifier and field key of the target allowed types. The last two
parameters are the same as the ones in the callback function. It will
find the `AllowedTypesMetadata.custom` for the passed in allowed types
and determine if the `incrementalSummaryHint` symbol is set to true.


[AB#41866](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41866)