Fix assert 0xb7b (#25735)
There's a corner case where processing incoming ops (join/leave ops in
particular) can trigger a connection state transition which can trigger
a flush (but there are no changes to flush), and assert 0xb7b fires - it
asserts that we should never flush while processing incoming ops. This
protects against reentrancy, which is not an issue in this case.

The fix is to only assert if we have something to flush.