feat(counter): Add rollback support for SharedCounter (#25771)
## Description

This PR adds rollback support for the `SharedCounter` DDS.

## Implementation

`SharedCounter` only has one type of op (`increment`), which will be
reflected optimistically by incrementing the counter's value. To
rollback we can simply call `incrementCore()` with opposite of the value
that it was optimistically incremented with. This will also emit an
`incremented` event with the amount that was rolled back and the new
value of the counter.

This PR also adds additional bookkeeping for pending ops to ensure we
handle local ops properly. This may not be technically necessary but can
help catch bugs in the future and is more in line with how other DDSes
handle local ops.

## Testing

This PR adds the following testing:
- UTs for Counter rollback scenarios
- Enables rollback for existing Counter fuzz tests
- Counter is now used in local server stress tests

## Misc


[AB#44705](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/44705)