(compat) Added a new error type for layer incompatibility error (#25784)
Follow up to #25706.

Added a new error type `layerIncompatibilityError` which will be used
when a layer incompatibility is detected between Fluid layers. This will
help applications detect when an error is because of layer
incompatibility and take appropriate steps.

The new error type is added as an legacy / alpha API. Since the
`FluidErrorTypes` type is legacy / beta, added a `FluidErrorTypesAlpha`
type where new error types will be added to stage them. Once the
breaking change window has passed, error types will be moved to
`FluidErrorTypes`.


[AB#51811](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/51811)