fix(test-drivers): use valid URL for hostUrl in docker config (#25837)
## Description

In PR #25610, the `hostUrl` in the docker config was set to `'this value
is unused'` with a comment indicating it didn't seem to matter. However,
it turns out that the URL does need to be a valid URL.

This change fixes the issue by setting `hostUrl` to `http://localhost:3003`,
which matches the alfred service URL (same as `ordererUrl`). Alfred is
the main service endpoint in the docker setup, accessible via the nginx
proxy on port 3003.

## Changes
- Updated `hostUrl` from `'this value is unused'` to
`http://localhost:3003` in the docker config
- Removed the misleading comment about the value being unused