(tree) Add version FieldBatchFormatVersion.v2 for FieldBatchCodec (#25811)
## Description
Added a new version `v2` to `FieldBatchFormatVersion`. This new version
will be used for `FieldBatchCodec` when `minVersionForCollab` is >=
2.73.0. Incremental summaries will be enabled for
`FieldBatchFormatVersion.v2` and later - a follow up PR will make
necessary changes to forest summarizer for this. This change adds
validation that the `FieldBatchFormatVersion` is v2 when incremental
encoding / decoding is used.

Added stronger typing `FieldBatchFormatVersion` in places where number
was used for field batch codec versions.

Updated the codec snapshot tests to run for `FieldBatchFormatVersion`
versions v1 and v2. Added smoke tests for `FieldBatchFormatVersion` v2.