Replace tree-agent editor binding with direct context creation (#25872)
Editors now receive trees as input rather than context. The
implementation of the editor is expected to call a (newly exposed)
`createContext(tree)` function to get the context for the executing
code. This change makes the input to an editor fully serializable - both
the tree state and the code string are serializable. This was not the
case before since the context object was not serializable. Having editor
inputs be serializable allows them to be run in a different process than
the process owning the `SharedTreeSemanticAgent`, which enables
additional scenarios for sandboxing (e.g. running tree-agent in process
A but actually executing the generated code in sandboxed process B).

Additionally:
* Exposed the Context type
* Parameterized various types (e.g. the editors) over a TSchema
* Removed `UnsafeUnknownSchema` from type signatures
* Exposed `viewOrTree` on `Subtree` so that we can go back to a
`ViewOrTree` from a `Subtree`
* Narrowed input type of the prompt function so it is no longer
invariant over TSchema (and can't have anything passed in without a
cast)
* Unexposed the default editor
* Added ViewOrTree type since the "view or tree" pattern is now in
multiple places in the user-facing API surface.
* Replaced binding unit tests with unit tests using custom editors.
* Various cleanup to docs and code
* Fixed bad string in tree-agent-ses test.