Centralize test exception validation logic (#25882)
## Description

Tree's exception validation test logic was nice enough it got copies
twice into other places, which were then not updated as it was improved.
This is silly. Instead of duplication, we can have code sharing. This
deduplicates the code, and establishes a plan for a follow-up to update
validateAssertionError to match the cleaner API for validateUsageError.

experimental/tree has been updated to the validateAssertionError2 API to
show what its usage looks like. Larger changes to more likely to
conflict parts of the codebase will follow if this is approved.