tree: remove unsupported codecs (#25893)
## Description

Removes reading and writing capabilities for the following formats:
* EditManager: v1, v2
* Message: vUndefined, v1, v2
* SharedTreeChangeFormatVersion:  v1, v2
* ModularChangeFormatVersion: v1, v2
* OptionalChangeset: v1
* SequenceChangeset: v1

Removing these formats reduces the package size and reduces the burden
of maintenance when introducing new in-memory representations.

## Breaking Changes

* Removes `SharedTreeFormatVersion` (exposed as alpha). It no longer
serves a purpose (and should have been removed with
https://github.com/microsoft/FluidFramework/pull/25752)
* Summaries and ops with `SharedTreeFormatVersion.v1` or
`SharedTreeFormatVersion.v2` configurations are no longer loadable (a
usage error explaining as much will be thrown). This is not expected to
cause issues because such formats were only used by default prior to
releasing FF 2.0 and have only been usable since through alpha APIs. We
do not know of any customer (1st or 3rd party) that depends on such
formats.