Simple to stored fixes and cleanup (#25899)
## Description

The new alpha APIs for handling simple schema as stored schema did not
work correctly for staged schema when the simple schema was not
implemented by a full view schema.
[AB#53901](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/53901)

Now that all the information needed to translate to stored schema is
present on simple schema (Including the SchemUpgrade toke added in this
change), it is possible to deduplicate the two code paths for simple
schema to stored schema. This change does this deduplication, and also
fixes the remaining code path to handle staged schema correctly.

Additionally, much more testing for this has been added.