(compat) Updated build-tools to check that layer generation is updated during release (#25873)
Added a new check function that will run during a release. For minor and
patch releases, it will check the generation for layer compat has been
updated before releasing and bumping versions. The check function runs
the `pnpm run -r layerGeneration:gen` command and if there are any
changes, prompts to commit them before proceeding with the release. This
is similar to how type tests and release notes are updated.

Follow up to https://github.com/microsoft/FluidFramework/pull/25670 and
https://github.com/microsoft/FluidFramework/pull/25835.


[AB#51927](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/51927)