fix(eslint-plugin-fluid): `no-hyphen-after-jsdoc-tag` rule (#25923)
The RegEx used to implement this rule was overly permissive and yielded
lots of false positives. This PR updates the rule to leverage the TSDoc
parser directly to extract block tags/comments and evaluate them
explicitly.

Unfortunately, this makes it harder to implement auto-fix behavior, so
that has been removed.


[AB#50920](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/50920)