fix(api-markdown-documenter): Fixed an issue where `MarkdownRenderer` would emit trailing `\` characters when processing certain summary comments. (#25933)
The issue here is somewhat complicated. The following is an attempt to
provide a high-level-ish overview of what's going on.

## The Problem

Keeping the description at a high-level, `CommonMark` specifies 2 ways
to force an explicit line break: either ending a line with 2 spaces or
ending a line with a backslash.
- https://commonmark.org/help/tutorial/03-paragraphs.html

`mdast-util-to-markdown`, the library we use to emit prefers the latter,
as it is more explicit.

In most cases, this is irrelevant, and Markdown parsers that adhere to
`CommonMark` should tolerate this. Docusaurus doesn't appear to, causing
spurious trailing backslash characters to appear in *some* of our API
documentation. E.g.
https://fluidframework.com/docs/api/fluid-framework/allowedtypesfullevaluated-typealias

But our library shouldn't need to output forced line breaks. We build
our Markdown AST trees manually, and so we control the output structure.
We generally avoid explicit line breaks, and instead use `Paragraph`
nodes when separating blocks of text.

But for contents parsed by `TSDoc`, we _transform_ their parsed content,
making certain assumptions about how the parser behaves.

The case that was causing problems for us is a case where `TSDoc` fails
to trim trailing whitespace from comment blocks (generally this
manifests in the summary component). In particular, when a summary
comment is followed by 2 or more modifier tags on a subsequent line,
TSDoc fails to trim the summary comment, yielding something with
trailing explicit line breaks.

### Example

Given the following comment:

```typescript 
/**
 * I am a comment
 * 
 */
```

The TSDoc parser will yield a summary block of `I am a comment`

But given the following comment:

```typescript
/**
 * I am a comment
 * 
 * @sealed @public
 */
```

The TSDoc parser will yield a summary block of `I am a comment \n`

## The Solution

The solution is to better trim paragraph contents given to us from the
`TSDoc` parser. Previously, we would trim any leading/trailing line
breaks we encountered. Now, we also trim and leading/trailing whitespace
content.

This simplifies the Paragraph trees we hand off to
`mdast-util-to-markdown`, removing the need for them to inject trailing
`\` characters to force line breaks.


[AB#53737](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/53737)