(tree) Fixed incremental summary not working for map and record nodes (#25946)
Incremental summary / encoding is not working for map and record nodes
even if they are opted in via the schema. This is because the encoder
checks for the opt-in for object nodes only in `getNodeEncoder`.
This change fixes that by checking if incremental encoding is opted in
for map nodes and record nodes (which are also map nodes internally).
Also, updated the tests to validate that all node kinds are correctly
opted into incremental summary.


[AB#54413](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/54413)