(test) Added support for testing compat versions of packages not available in older versions (#25915)
Currently, the test infra installs packages for older versions (than the
current) that were available in all the older versions that the test
supports, going back to 1.4.0.
This means that we don't get compat testing for packages that are not
available in these older versions. For example, the
"@fluidframework/tree" packages is available from version 2.0.0 onwards.
We would still want compat coverage for SharedTree - layer compat and
cross-client compat with versions that are 2.0.0+.

This change adds support for testing compat for these packages. The
basic idea is to have compat testing going back to the version that
these packages were first available:
1. In the `testApi.ts` file, added a `minVersion` to each package that
is installed by the test infra. The infra will only install the package
if the requested version is >= the `minVersion` for that package. For
`@fluidframework/tree` package, the `minVersion` is 2.0.0.
2. In the `CompatApis` exposed by the test infra, the entry
corresponding to packages that are not installed will be undefined. It
is the responsibility of the tests to ensure that it only runs in valid
compat configurations. For example, the `packages:tree` and `SharedTree`
properties of `dataRuntimeApi` will be undefined in version `1.4.0`. The
tests in `treeCompat.spec.ts` skips running the cross-client compat
tests in this configuration.
3. Added a `mode` property to `CompatApis` which tells whether the test
is running in layer compat, cross-client compat or no compat mode. This
will help tests skip running in certain configurations. The tests in
`treeCompat.spec.ts` uses this to skip running in cross-client compat
mode when one of the client versions is less than 2.0.0.


[AB#54447](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/54447)