fix(intervalCollection): Fix bugs seen from oracle (#25961)
This PR fixes two bugs which were seen in the oracle testing for
interval collection.

1. Missing `deleteInterval` event when intervals slide off during
rebasing

When an interval slides off the string during rebasing the code was
calling `removeExistingInterval` which removes the interval without
emitting a `deleteInterval` event. The oracle still thinks the interval
exists, causing validation failures.

2. Events emitted for deleted intervals

When an interval has been deleted `latestInterval === undefined`, but
remote ops are still being acknowledged, the code emits events with
undefined as the interval, violating the API contract that says
`interval` must never be undefined.