Enable undo/redo delete Shared Array ops in fuzz tests and fix rollback bug (#25980)
This PR enables toggling a Shared Array delete op during our fuzz tests.
It achieves it by adding the delete ids list to the possible ops to be
toggled while generating fuzzed ops.
It also fixes an issue when toggling an op that was previously
rollbacked. An op that was previously rollbacked should not be toggled
after that, since rollback ops are meant to behave as if they never
existed. Applications could be trying to toggle such ops given that
every shared array op emits a "revertible" event which applications can
listen to and generate a list of ops that could potentially be reverted.
The solution was adding an optional attribute to the shared array entry
to be able to recognize if such array entry was rollbacked and skip
toggling if that was the case.
After solving that bug, another seed started failing with a different
problem related to rollback while in attaching state that will be
tracked and solved separately.