(tree) Add versioning to shared tree's summary (#25876)
## Description
The summaries written by Shared Tree and its summarizables are not
versioned. Any changes to the summary format can lead to failures in
clients that do not understand this new format. For example, the
incremental summarization feature changes the forest's summary format by
adding more nodes to its summary tree which older clients won't be able
to read. Making this change behind a new version makes it safer as we
can fail fast on seeing unrecognizable versions.
This change adds versioning by adding a metadata blob to the summary
tree of the shared tree and its summarizables. This metadata blob will
contain the format version of the summary. Every time the format of the
summary changes, a new version should be added so that clients that
don't understand this format will fail.

Here's how versioning will be added:
1. The current summaries generation (before this PR) are considered
format version 1.
2. WIth this PR, a metadata blob is written into the summary and thus,
changing the summary format. This summary's format version is 2. This is
fully backwards compatible with version 1 and will not break older
clients.
3. In a future change, the summariables will write format version 3
based on the minVersionForCollab. This version will change the format
such that older clients that read / write version 1 and 2 will fail on
reading version 3. This will be done to ensure that summary format
changes in the future don't silently fail at random places.
4. Any changes to the summary format by a summarizable after this point
should do so with a new format version.


[AB#53723](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/53723)