fix(tree): commit enrichment after transaction abort (#25978)
## Description

Aborting a transaction used to put the tree in a state that would
trigger an assert when sending some undo/redo edits to peers.
This would prevent some undo/redo edits from being sent and would put
the tree in a broken state that prevented any further edits.
This issue could not have caused document corruption, so reopening the
document was a possible remedy.

After this PR, aborting a transaction no longer puts the tree in such a
state, so it is safe to perform undo/redo edits after that.

This bug was due to the fact that `TreeCheckout` overwrote its
`DetachedFieldIndex` member with a new instance when performing a
transaction rollback but had no way of informing the
`SharedTreeReadonlyChangeEnricher` that it should now be reading from
the new instance during change enrichment.

This PR fixes the bug by ensuring that a given `TreeCheckout` instance
always uses the same `DetachedFieldIndex` instance. This requires
introducing a way of snapshotting and (on transaction rollback)
restoring the `DetachedFieldIndex` state that does not require replacing
the instance.

This PR also reduces the `DetachedFieldIndex` API surface made available
to `SharedTreeReadonlyChangeEnricher` by introducing the
`ReadOnlyDetachedFieldIndex` interface. This is not required as part of
this change and could be done in a separate PR.

## Breaking Changes

None

## Reviewer Guidance

Let me know if you think it's worth spinning up a separate PR for the
introduction of `ReadOnlyDetachedFieldIndex`.