(compat) Incremenent layer compat generation for Fluid layers to 3 (#26029)
This change updates the `generation` for the Fluid layers to 3. This
should be updated every month. The last `generation` 2 was released in
[2.72.0](https://github.com/microsoft/FluidFramework/releases/tag/client_v2.72.0)
which was more than a month ago.

Note: The logic to update this automatically is in but it requires a
change in build-tools. This PR is in progress -
https://github.com/microsoft/FluidFramework/pull/26015. But we have
decided to wait for the next release to complete this PR because there
may not be enough time to merge that PR, release build-tools and
integrate that in to the main Fluid repo.