fix(container-runtime): Distinguish "Duplicate DataStore Id" errors between permanently corrupted op stream v. single-client one-time failure (#26021)
Today we throw a DataCorruption error with message "Duplicate DataStore
created with existing id" if we process a DataStore attach op but
already have that ID (not counting when it's our own attach op).

However, if we have that ID for an unbound DataStore - meaning, we
created it locally and it doesn't exist anywhere else - then as long as
we crash the container without trying to attach our copy, this case will
not result in a corrupted container.

We have seen this play out in cases where a partner had a bug that
violated protocol and returned an existing container when the client
thought it was creating a new container.

In that case (or any such case), we want to properly represent the
impact of this error case - it's not permanent DataCorruption,
subsequent sessions succeed.