Updated ClientVersionDispatchingCodecBuilder (#26011)
## Description

Proposed design for how we should handle codec version selection.

This cuts down on per codec logic a lot, generating as much as possible
from a single declarative specification of what codecs are available.

It also matches the runtime configuration approach of providing a way to
override codecs (in this case by name) to force specific versions. We
currently do not expose the name of any codec in a stable way so this
can really only be used experimentally.

There is still a lot of work to do to migrate and refactor more cases
before the old utilities can be fully removed, but this seems like a
good incremental step.