refactor(tree): refactor SquashingTransactionStack (#26053)
## Description

Changes `TransactionStack` so that `onPush` callbacks can (in addition
to returning a callback for the matching pop) return a new `onPush`
callback to be invoked for nested transactions.

This change is leveraged by `SquashingTransactionStack` to do two
things:
1. Separate the logic associated with inner transactions from the logic
associated with outer transactions. This removes the need for repeated
checking of `this.isInProgress()`.
2. Allow the logic for both inner and outer transactions to refer to
state that the outer `onPush` logic sets up without having to assert.

These changes pave the way for future changes which will introduce more
transaction state.

## Breaking Changes

None