(compat) Update Fluid compat layers to use auto-generated generation value from client-utils (#26072)
The `generation` value for layer compatibility is auto-generated in the
"client-utils" package. All the compat layers have been updated use this
generation value instead of a hard coded value.

Also, ran the `layerGeneration:gen` command in the `client-utils`
package to tag the `generation` value in the `layerGenerationState.ts`
as `@internal`.


[AB#55196](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/55196)