(compat) Validate layer compat only for layers that expose layer compat state  (#26075)
Updated the layer compatibility validation to run only for layers with
versions that has the layer enforcement logic.
The exception to this is when the Runtime validates the Loader layer
(details below).

This will ensure that if some of the layers have custom implementations
of objects exposing layer compat state, they don't start failing with
layer incompatibility errors. For example, some customers have custom
implementations for data store runtime for use as placeholders where
their data store factories are removed. Those will start failing with
the current implementation.

The only exception to this is when the Runtime layer validates whether
Loader layer is compatible. In this case, if the Loader layer doesn't
expose layer compat details, a default one will be used which basically
treats the layer as if it's generation was just before layer compat was
introduced. The reason for this is because this layer boundary is the
one that can drift apart and we want customers to be able to identify
these scenarios and fix them at least initially while we roll enabling
the validation. We can remove this exception once the feature is fully
rolled out and working as expected.
As a safety measure, added a feature flag that be used to disable this
exception.

Also, added a feature flag that can be used to disable layer
compatibility validation altogether. Thsi is another safety mechanism to
provide an escape valve when initially rolling out this feature.

When layer incompatibility failure is bypassed due to the above, a
telemetry "LayerIncompatibilityDetectedButBypassed" is logged. It is
sampled to only log once per session per layer combination.


[AB#34850](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/34850)